+++ apps/nextjs-website/react-components/components/index.ts
@@ -1,0 +1,3 @@
+import Hero from './Hero/Hero';
+
+export { Hero };
