--- apps/nextjs-website/react-components/index.ts
@@ -1,3 +1,0 @@
-import Hero from './components/Hero/Hero';
-
-export { Hero };
