+++ apps/nextjs-website/react-components/types/index.ts
@@ -1,0 +1,3 @@
+import { HeroProps } from './Hero/Hero.types';
+
+export type { HeroProps };
