+++ apps/nextjs-website/tsconfig.json
@@ -17,7 +17,6 @@
     ],
     "paths": {
       "@/*": ["./src/*"],
-      "@react-components": ["./react-components"],
       "@react-components/*": ["./react-components/*"],
     }
   },
